feat: Improve the data compression by serialising plural form objects to strings
BREAKING CHANGE: Plural data created by this version of the library cannot be consumed by *older versions*. It would be unusual, if you created data by tis version and fed them to an older version of the library, but nevertheless, be aware of it. If you create data by this version, use them with the same major version or a newer one, as long as it is documented to be feasible.

Plural data created by older versios of this library can still be used with this version. The code is beckwards compatible.